[Substratum] Themes: SystemUI: Expose tint color for power menu icons
As of Android 11, items in the layouts folder can no longer be modified.
Therefore, this commit exposes the tint color of the icons in the power
menu (except for the emergency button, which already has special tints).

Note that Android 14 always uses a dark background dialog for the power
menu in both light and dark modes, so the default color is the same as the
stock AOSP icons. This is a change from Android 12 and 13. .